Update NetworkController: network -> networkStatus
The `network` store of the network controller holds a mixture of two
things. It roughly tracks whether we have enough information to make
requests to the network and whether the network is capable of receiving
requests, but it also stores the version of the network (as obtained via
`net_version`).

Generally we shouldn't be using the network version for anything. It has
been completely replaced by chain ID, which all custom RPC endpoints
have been required to support for over a year now. So we really just
want to keep track of the network status.

This commit replaces the `networkStore` inside of NetworkController with
`networkStatusStore`, and changes how it is used and set. Previously it
would only get set to a network version once we obtained that, but since
we don't do that anymore, we will always set it to "active" unless no
chain ID is available for the network in question. (This holds even if
we are unable to determine whether the network supports EIP-1559.)